review(sync.test): more ulong cascades, SaturatingSub in BuildNewBatch
PeerInfoAllocationTests.BuildHeaders: drop (ulong)number cast (new int WithNumber overload).

SynchronizerTests.AddBlocksUpTo / AddHighDifficultyBlocksUpTo: int → ulong loop counters
and parameters; drop the (long)block.Number cast.

ReceiptsSyncFeedTests.Scenario: lift `(int)_pivotNumber` into a single local so the
loop bodies don't repeat the cast on every comparison.

E2ESyncTests.HeadPivotDistance / PartialBalSyncHeadPivotDistance: int → ulong consts;
SetPivot/SyncUntilFinished params ulong; SaturatingSub for the head-finalized gap.

HeadersSyncFeed.BuildNewBatch: SaturatingSub replaces the explicit
`_lowestRequestedHeaderNumber >= requestSizeU ? ... : 0UL` clamp.

+++ src/Nethermind/Nethermind.Synchronization.Test/PeerInfoAllocationTests.cs
@@ -113,7 +113,7 @@ namespace Nethermind.Synchronization.Test
             ArrayPoolList<BlockHeader> headers = new(end - start + 1);
             for (int number = start; number <= end; number++)
             {
-                headers.Add(Build.A.BlockHeader.WithNumber((ulong)number).TestObject);
+                headers.Add(Build.A.BlockHeader.WithNumber(number).TestObject);
             }
 
             return headers;
